Allow building std.ndll as a static library.
In order to ensure that all exported functions are included
into the executable, this commit adds a script that collects
a list of such functions from the library sources and creates
a table listing them all.

Afterwards, this table is used in a function that initializes
the library, thus providing explicit dependencies for the linker.
The linking executable is supposed to call neko_stdlib_init.